Add contents to fried-chicken.html file
Add an image of the finished dish under a h1 heading of the recipe's name.
Under the image include a description of the recipe. Under the description
add an 'Ingredients' section listing the needed ingredients. Finally, under
the ingredients section, add a 'Steps' section detailing the steps to the recipe.